Refactor TextBox
